Adding puppet-usm module
Unified Software Management (USM) needs a puppet module
to setup its authenticated endpoints, ports, etc..

The puppet-usm is not explicitly added to the ISO.
It gets pulled in through the puppet-manifests dependency.

Note: this follows the same algorithm and code paths that
are used for setting up patching through puppet.

The USM ports are:
 API port (private): 5493
 Controller port:    5494
 Agent port:         5495
 API port (auth):    5497
 Admin port (auth):  5498

Default URLS: (pre bootstrap these are not in keystone)
 Private URL http://127.0.0.1:5497/v1
 Public  URL http://127.0.0.1:15497/v1
 Admin   URL http://127.0.0.1:5497/v1

After bootstrap (DC mode) endpoints look like
 SystemController internal  http://192.168.204.2:25497/
 SystemController public    http://10.10.10.2:25497/
 SystemController admin     https://192.168.204.2:25498/

 RegionOne internal  http://192.168.204.2:5497
 RegionOne public    http://10.10.10.2:15497
 RegionOne admin     https://192.168.204.2:5498

Note: a sysinv change enables this puppet code and testing
was also done with that code.

Test Plan:
  PASS: Build/bootstrap/unlock AIO-SX
  PASS: Build/bootstrap/unlock STD (2 controllers)

Story: 2010676
Task: 48408
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I263c4ca135ac9b8bf4383b74dd44754c043b7802