Fix task ordering to prevent race condition on ptp config gen
Changed the task ordering for the ptpinstance runtime manifest. The
previous ordering could result in a race condition where collectd was
being restarted before puppet had finished writing all of the ptp
configs. This would result in collectd initializing and not reading all
of the ptp config files.

Changed the chaining arrow to ensure that collectd is not restarted
until the config generation has completed.

Test Plan:
PASS: Generate multiple ptp configs and run the manifest with system
ptp-instance-apply. Ensure that collectd starts after config generation
is complete. Performed dozens of iterations over the course of other
development work and the race condition was not observed again.

Story: 2010723
Task: 48442
Change-Id: I057c80a1c8488488ba26c739c9083ebfe63738d7