Request runtime configuration in case of empty config info
During patch apply tests, it was detected that if all the firewall
puppet classes are not provided during boot, we need to mark the
filesystem with .platform_firewall_config_required in /etc/platform/
so sysinv-agent can request a runtime execution to conductor. The
audit operation already looks for this file, this change just adds
an extra condition.

Since this can happen also in the worker nodes, this logic is added
there during script execution. The worker script itself will not
declare error, just creates the flag to execute the runtime config.

In all test scenarios the validation was done with the verification
that the firewall was created after patch application.

Test Plan:
[PASS] Apply/Remove the firewall feature as a patch in AIO-SX
[PASS] Apply/Remove the firewall feature as a patch in AIO-DX
[PASS] Apply/Remove the firewall feature as a patch in Standard

Story: 2010591
Task: 48461

Change-Id: Id9a22ff11cd347cf7a162662b7d18d9955206411
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>