Fixed puppet error when installing a worker node
The puppet error is due to the resource relation specified
using ordering arrow (->). Deleted the relation as the
AppArmor permission change of /etc/apparmor.d is not related
to the previous block.

Test Plan:
PASSED: Standard: Boot with the ISO and check if 'sys_protected:rwx'
        permission is added for /etc/apparmor.d directory.
PASSED: Standard: Check for puppet error "Evaluation Error: Illegal
        relationship operand" on controller and worker nodes.
        There should not be any such error.

Closes-Bug: 2026666

Change-Id: I6dd3e461be0e24844fff0ca240b6db7721fa1801
Signed-off-by: Jagatguru Prasad Mishra <jagatguruprasad.mishra@windriver.com>