Disable admin network failover behaviour
A requirement for a subcloud's admin network is that its
subnet information be able to be updated at runtime without
host lock / unlock.  In an AIO-DX system this can cause
issues in certain circumstances as the disablement /
enablement must be done via puppet and is affected
by the ordering a user performs each action as well as
the currently running manifests on each host.

A corresponding SM commit will disable the failover
behaviour for the admin network, while still maintaining
the alarming.

This commit eliminates the ifdown/ifup behaviour for the
admin network when it is disabled / enabled / updated, as
it is not needed anymore.

Test plan:

- No uncontrolled swacts while re-configuring admin subnets
  or reverting to the management subnet (deleting the admin
  address pool)

- Alarms still generated on interface down / heartbeat loss

- Switching back and forth between admin network / mgmt
  network via dcmanager.

Story: 2010319
Task: 47707

Change-Id: I5a5f2919cb5a815c0e6dbf8442458a27500cee14
Signed-off-by: Steven Webster <steven.webster@windriver.com>