Select correct config order for ifup if bonding is present
It was detected that if the bond interface name starts with a letter
after 'e' (the usual letter for ethernet interfaces) it will be placed
in a position in the /etc/network/interfaces.d/auto file that will
prevent the correct bonding configuration.

This change creates a order in the auto file as
"lo [bond interfaces] [ethernet interfaces] [vlan interfaces]"
as the ifenslave package also configures the slave interfaces when
the bonding is configured.

Test Plan
[PASS] configure a bond named "ae0" and observe correct configuration
       after lock/unlock
[PASS] configure a bond named "pxeboot0" and observe the correct
       configuration after lock/unlock
[PASS] on top of pxeboot0 add vlan interfaces and observe the correct
       configuration after lock/unlock
[PASS] remove content in /etc/network/interface.d/ and execute a
       lock/unlock cycle to test apply_network_config.sh
[PASS] add/remove static routes in runtime execution and after a
       lock/unlock cycle

Closes-Bug: 2028877

Change-Id: I55625cb64e815899f6fb33759b8018c26cd52de2
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>