Add use_usm parameters to dcmanager.conf
Add use_usm parameter to /etc/dcmanager/dcmanager.conf to be used
in a switch to choose between the usm and patching/sysinv api. Once the usm is fully integrated, this parameter will be removed.

Test Case:
Apply the change and restart the system controller and
ensure use_usm parameter is in /etc/dcmanager/dcmanager.conf

Story: 2010676
Task: 48530

Change-Id: I3cda0617c5b38edc82668d7f508e9f4876640bc7
Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>