In IPv6 only search addresses in tentative if interface is UP
During usage it was observed, in IPv6, that if a interface is with
operational state in DOWN, it will take more time to leave the DAD
state "tentative", if the system starts without carrier the flag
remains until a carrier signal is provided.

Since interfaces in DOWN are not affected by duplicated address
in the network, this change adds a script to test this condition
over active interfaces only.

Test Plan:
[PASS] remove the signal of an ethernet port that is configured and
        execute lock/unlock. Observe that there will be no puppet
        error in "wait-for-tentative".

Closes-Bug: 2030960

Change-Id: I8befd232324a2386879627f0d19ea7d14cccc126
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>