Conditionally deprovision guest-agent service
After the Debian migration it was noticed that the guest-agent service
was causing issues when enabled (i.e., when stx-openstack is applied)
and was actually not being used on CentOS for a while.
The decision was to completely remove it [1] and puppet was cleaned to
not have mentions to this service [2].

Previous changes work fine for newly installed systems, but if one needs
to patch or upgrade a system that has this service enabled, puppet needs
to enforce this deprovision before any sm-db migration/update is done.

[1] https://review.opendev.org/c/starlingx/ha/+/870859
[2] https://review.opendev.org/c/starlingx/stx-puppet/+/869474

TEST PLAN:
PASS - build-pkgs -c -p puppet-manifests
PASS - build-image
PASS - Bootstrap AIO-SX (virtual)
PASS - Check puppet.log to see if guest-agent was deprovisioned

Story: 2010317
Task: 48598

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: I66c8733c48d0f87d1394901d427e94d913dbcca8