Add use_usm parameter to dcorch.conf
Add use_usm parameter to /etc/dcorch/dcorch.conf to be used
in a switch to choose between the usm and patching api. Once the usm
is fully integrated, this parameter will be removed.

Test Case:
Apply the change and restart the system controller and
ensure use_usm parameter is in /etc/dcorch/dcorch.conf

Story: 2010676
Task: 48610

Change-Id: I9ff6e3a7c5eaef81e0d6c9ca7529fe4c278275bc
Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>