Set default value for private_dc_ip_address
It was found that when applying a patch containing the code for
enabling the admin network on distributed cloud systems, a
problem could occur after unlock because of an undefined value
for 'private_dc_ip_address'.

This is because the value of this puppet parameter is only set
in the system.yaml hieradata, which is not written on normal
host lock/unlock.

The value is generated however, when a user assigns an interface
to the admin network, or performs another action that leads to
the system hieradata being generated.

If the user has not assigned an interface to the admin network,
or if the user has no interest in using the admin network, this
commit ensures that the value of the private_dc_ip_address is
that of the private_ip_address (ie. mgmt).

Setting the private_dc_ip_address to the private_ip_address
as default ensures that every component using the haproxy::params
behaves as it did pre-patch.

Testing:

Success path:

- Install, lock / unlock AIO-SX system (stx 8.0 + patch)
- Install, lock / unlock an AIO-DX DC subcloud (stx 8.0 + patch)
- Update a DC subcloud to use the admin network after the system
  has been patched.

Regression:

- Install, lock / unlock AIO-SX system (dev build)
- Install, lock / unlock an AIO-DX DC subcloud (dev build)
- Install a DC subcloud using the mgmt. network (dev build)
- Install a DC subcloud using the admin network (dev build)

Story: 2010319
Task: 46911

Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: Ice16a4a3a2d9faa461c1c0f98d1ea0b0c7ce751a