Update kubelet system overrides on unlock
Add logic to the platform::kubernetes::configuration method
to generate the kubelet's systemd override file. This
change ensures the file is generated every time a host is
unlocked. This facilitates delivery of systemd service changes
via patches to existing installs.

This change is needed by bug 2027810 to ensure that the
orphan volume cleanup script is executed as part of the systemd
ExecStartPre kubelet service override.

This bug is an update for the this reverted commit:
https://review.opendev.org/c/starlingx/stx-puppet/+/895364

Test Plan:

Pass:  - Update the kube-stx-override.conf.erb file
       - Lock the AIO-SX host
       - Unlock the AIO-SX host
       - Verify that kube-stx-override.conf has been updated
       - Verify AIO-SX fresh install
       - Verify Standard Duplex lock/unlock and
         verify that kube-stx-override.conf has been updated
       - Verify Standard Duplex Install

Partial-Bug: 2027810
Change-Id: I4e47bce634c21396acb2e5f1540cac0be3ed34ec
Signed-off-by: Gleb Aronsky <gleb.aronsky@windriver.com>