Remove worker remote firewall scripts
The implementation for worker firewall avoided using local kubectl
commands. This required access to the keyring for remote ansible
ad-hoc commands and leaves the /opt/platform/.config mounted on the
worker.

Use kubectl command with /etc/kubernetes/kubelet.conf instead, so we
can refrain from mounting /opt/platform/.config

Since all firewall data is generated in the host's hierada file, the
worker node needs to be able to access the calico firewall resources.
To achieve that a ClusterRole and ClusterRoleBinding are added, via
the controller node, allowing access to only the necessary resources.

Test Plan:
[PASS] Install a standard setup and validate the worker node firewall
        configuration
[PASS] Execute a DOR test in the cluster and check if the worker nodes
        install the firewall GNP and HE
[PASS] Execute worker node lock/unlock and check if the worker nodes
        install the firewall GNP and HE

Closes-Bug: 2038550

Change-Id: Icf31b513427120fe81c53be21b8d8a81a8e323f8
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>