Add k8s cfg file to the OAM firewall script
In the change
https://review.opendev.org/c/starlingx/stx-puppet/+/897467 the OAM
firewall was not updated to pass the k8s config file as argument
to calico_firewall_apply_policy.sh. It then created an error that
prevented the global network policy to be created, making the OAM
interface to block all traffic, except for the failsafed ones.

This change corrects that

Test Plan
[PASS] In AIO-DX remove the current OAM GNP and execute lock/unlock
        on one of the controllers, verify the OAM GNP is recreated.
[PASS] In AIO-DX remove the current OAM GNP and force the runtime
        execution by creating the file
        /etc/platform/.platform_firewall_config_required and observe
        the request to recreate the OAM GNP

Closes-Bug: 2038550


Change-Id: Ica03dbf6ffd9f6f592fa53efa40293191203377a
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>