Persist iSCSI Initiator Name, NVMe hostid and hostnqn [2/2]
This is a part of solution to persist some iSCSI and NVMe parameters
across host reinstallations, host upgrades and backup/restore.

To support Dell Storage applications, it's needed to generate a
different iSCSI Initiator Name, NVMe hostid and hostnqn on each node.

We will persist this information in the i_host table of the sysinv
database in the following columns:
  - iscsi_initiator_name
  - nvme_host_id
  - nvme_host_nqn

We invoke the classes in the unlock stage just to overwrite and
restart the service once, we chose this location because if we
run it at runtime(config::hosts) the file will be overwritten
some times and the services restarted together and this may cause
a future issue in other scenarios that use iscsi service.

Test Plan:
  PASS: Node parameters preserved after host reinstall
  PASS: Parameters preserved after backup/restore
  PASS: After AIO-DX fresh install, each node has an unique
      iSCSI initiator name, NVMe hostid and hostnqn paramenter

Story: 2010693
Task: 48962

Depends-On: https://review.opendev.org/c/starlingx/config/+/898861

Change-Id: I28c384f152519c7b5dae701d951420da38e1acf3
Signed-off-by: Hediberto Cavalcante da Silva <hediberto.cavalcantedasilva@windriver.com>
Signed-off-by: Tiago Leal <Tiago.Leal@windriver.com>
Signed-off-by: Gustavo Ornaghi Antunes <gustavo.ornaghiantunes@windriver.com>