Update ceph-mon ip due to MGMT reconfiguration on AIO-SX
During a management network reconfiguration on AIO-SX with Ceph
storage backend, it is necessary to update the ceph-monitor to
the new network.

This commit verifies if the '.ceph_mon_reconfig_required' flag
exists, and if so it is used in the mandatory controller-0 unlock
during the mgmt reconfiguration by the "platform::ceph::monitor"
puppet class, which will proceed with the ip update.

Test Plan:
 PASS: Reconfigure mgmt network on AIO-SX ipv4 with Ceph backend
 PASS: Reconfigure mgmt network on AIO-SX ipv6 with Ceph backend
 PASS: After controller-0 is unlocked, check if Ceph is healthy with
       'ceph -s'
 PASS: Reconfigure mgmt network on AIO-SX ipv4 without Ceph backend

Story: 2010722
Task: 48974

Depends-on: https://review.opendev.org/c/starlingx/config/+/898932

Change-Id: Ie3b46e1a9577a9193cad85af1eb3b95e88a10857
Signed-off-by: Gabriel de Araújo Cabral <gabriel.cabral@windriver.com>