Fix unmask service conditional for isolcpu_plugin
After a kubelet upgrade, when restarting the isolcpu_plugin,
the Puppet code needs to check that the isolcpu_plugin has
been explicitly masked and not disabled. A disabled setting
would indicate that the isolcpu_plugin should not run on that
particular node, unlike a masked isolcpu_plugin. This change
ensures that we differentiate between isolcpu_plugin being masked
vs disabled, which in the former we would want to start.

This commit corrects an erroneous update to line 952
in commit https://review.opendev.org/c/starlingx/stx-puppet/+/894545/9..10

Test Plan:
Pass: This code was tested was tested as part of Bug #2036985
      but was not merged in it's entirety.

Closes-Bug: 2041686
Change-Id: I5192ac515be3e3804ec77a14a6515ffe1da0d26a
Signed-off-by: Gleb Aronsky <gleb.aronsky@windriver.com>