Modify Docker Service to ensure file system dependency.
Add a dependency on var-lib-docker.mount to the After
and Requires section of [Unit] and do a daemon-reload.
This change ensures that docker is dependent on the
correct file system mounts to start up. This corrects
a failure scenario where the docker service starts up,
shows that it is running, but is in an unhealthy state
because it started before its filesystem mounted. This
fix ensures that other software running in an upgrade
scenario, the system is able to pull the required
images.

Test Plan:

PASS - BUILD ISO and test install on AIO-DX using
       Virtual Box. After upgrade controller-1 unlocks.
       Verify that the all files/folders present in
       /var/lib/docker AND ability to pull image
       without having to restart docker service.
       Also verify that the remainder of the lab upgrade
       and activation succeeds.

PASS - Build ISO and test install on AIO-DX on physical
       hardware. After controller-1 unlocks, ensure that
       all files/folders present in /var/lib/docker AND
       ability to pull image without having to restart
       docker service. Also verify that the remainder of
       the lab upgrade and activation succeeds.

PASS - Build ISO and test install on DC physical hardware,
       where central cloud is STANDARD config, and sub-clouds
       are AIO-SX.  Verify that the upgrade completes and
       activation is successful on all machines.

PASS - creation and application of service parameter:
       docker/proxy/http_proxy, docker/proxy/https_proxy,
       docker/proxy/no_proxy. Observed successful runtime manifest
       execution and create of http-proxy.conf files for containerd and
       docker.

Closes-Bug: 2043399

Change-Id: If4dd0c8426a218671a946ec54242fe3c20808714
Co-Authored-By: Robert Church <robert.church@windriver.com>
Signed-off-by: Joshua Reed <joshua.reed@windriver.com>