Fix up python script to change K8s control plane params
As documented here[1], if "kubeadm init" is called without explicitly
specifying the address for the apiserver to advertise, then it will use
the IP address that would be used by default to talk to the default gateway.

The change_k8s_control_plane_params.py script had a number of cases where
it called "kubeadm init" without explicitly specifying the address to
advertise, and this resulted in a bug where the default IP address was
used rather than the cluster host address.  This could cause problems in
a number of ways.  In some cases the default IP address at the time of a
"restore" operation was an auto-configured address that would not exist
after the "restore" was complete.  In other cases the default IP address
was the OAM address, but the end user can change that.  In either case
it could result in the kube-apiserver container failing its health checks
and being taken down by kubelet.

A separate problem was that in an attempt at fixing the above issue the
script explicitly modified the kubeadm-config ConfigMap to add the
"advertise-address=<address>" flag  to the kube-apiserver extraArgs field,
but specifying the controller-0 cluster host address.  This caused
kube-apiserver on controller-1 to incorrectly advertise the cluster host
address of controller-0.

The solution to both issues is to add an InitConfiguration section to
the kubeadm config file, setting the localAPIEndpoint field to the
cluster host address of the current host.

TEST PLAN:
PASS: Install AIO-DX, do backup and restore, ensure IP addresses used by
      kube-apiserver are correct.
PASS: Update kube-apiserver service parameter, ensure kube-apiserver is
      updated on both controllers and IP addresses remain correct.

[1] https://kubernetes.io/docs/reference/setup-tools/kubeadm/kubeadm-init/
[2] https://github.com/starlingx/stx-puppet/commit/2ceceb29

Closes-Bug: 2043537
Change-Id: Id6d4009722671a2b17acdd6bbadfc4049f2d4fac