Do not search for tentative addresses during AIO-SX upgrade
During upgrade the platform addresses are first configured
in the loopback to have them available during the bootstrap.

But later the ansible step "Apply network runtime manifest to populate
/etc/network/" runs and it activates the configured interfaces. This
leads to a situation were the controller address from a platform
interface may lead to a dad-failed event because the same address is
configured in the interface and loopback.

This change skips this verification during upgrade because the
following reasons:
1) during the upgrade bootstrap only the OAM interface is needed
for outside communications
2) The other platform interfaces will not be used in this stage
3) After bootstrap the system is automatically unlocked and in the
reboot return the tentative validation will take effect as the
regular network configuration will be applied.

Test Plan:
[PASS] AIO-SX upgrade from CentOS version to Debian
[PASS] Execute lock/unlock to test regular execution

Closes-Bug: 2044020

Change-Id: I1d61ef5260e3ddbfce8c3f01127a77a55524a746
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>