Filter correctly for VLAN in interfaces with class=data
It was detected that vlan interfaces with class data were not
correctly handled in the function is_vlan(). Since they do not use
name format used in ifupdown (vlanNNN or ethX.NNN) the test need to
include the pre-up operation that will create the device.

This change prevents that a non-starlingX interface with a network
configuration to be handled by apply_network_config. This script
handles only interfaces managed by starlingX, if there are other
interface configuration files in /etc/network/interfaces.d/ they need
to be handled by the originator (another process, the system
admin, others)

Test Plan:
[PASS] add a interface with class=data and type=vlan, then execute
       a) lock/unlock to create the interface in the node
       b) modify the MTU in the new interface
       c) lock/unlock to modify the interface in the node
       After that check interface creation in the kernel
[PASS] Create a vlan configuration file in /etc/network/interfaces.d/
       that is not part of starlingX's managed interfaces. Check that
       does not attempt to bringup the interface

Closes-Bug: 2045658

Change-Id: Ic10d83cd140e6a8283de8733c172142a9288b8c5
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>