Reduce network timeout setting in ldap.conf
NETWORK_TIMEOUT parameter specifies the amount of time to wait when
connecting to an LDAP server. If the server fails to respond in the
specified amount of time the client aborts the attempt.
Setting the NETWORK_TIMEOUT parameter, prevents long delays in
accessing the ldap server to get user information. The
delay can be caused by network connectivity fluctuations.
This commit reduces the NETWORK_TIMEOUT parameter value to 3s to
improve the duration of rehoming execution.

Test Plan:
PASS: In a DC system with a DX subcloud managed bootstraped and with
      in-sync status, remove the mgmt network between both systems and
      try to login the subcloud system. An ldap error should be raised
      and the terminal prompt should be available in less than 7sec.
PASS: In a DC system with a DX subcloud managed bootstraped and with
      in-sync status, remove the mgmt network between the systems and
      try to login the subcloud system and observe that ldap error is
      raised and terminal prompt is available in less than 7sec. Exit
      the system, add the mgmt network again and then login the
      subcloud system again. Observe that the login is completed
      successfully. An ldap error could be raised in less than 3sec due
      to the network connection being recovered and it is not presented
      after the network connection remains stable.
PASS: In a DC system with a DX subcloud managed bootstraped and with
      in-sync status, remove the mgmt network between the systems. SSH
      to subcloud system from a machine other than systemcontroller as
      sysadmin and issue a "sudo ls" command. An ldap error should be
      raised and the terminal prompt should be available in less than
      7sec.
PASS: In a DC system with a DX subcloud managed bootstraped and with
      in-sync status, remove the mgmt network between the systems and
      then add it again. SSH to subcloud system from systemcontroller
      as sysadmin via mgmt floating IP and issue a "sudo ls" command.
      Observe that the SSH login to subcloud and "sudo ls" commands are
      completed successfully. An ldap error could be raised in less
      than 3sec due to the network connection being recovered and it
      is not presented after the network connection remains stable.
PASS: Execute rehoming and observe that the duration has improved.

Closes-Bug: 2047950

Signed-off-by: Carmen Rata <carmen.rata@windriver.com>
Change-Id: I016342788b5e60c5e1e95ef9f885cddd7a34d60a